feat: added fetch tag categories function in TagsAdmin component
